+++ src/main/kotlin/step2/App.kt
@@ -1,6 +1,6 @@
 package step2
 
-import step2.config.ObjectAssemble
+import step2.assemble.ObjectAssemble
 
 fun main() {
     ObjectAssemble.calculatorView()
